added new function + cleanup
